Refactored customers and products

+++ src/DBAflevering/src/main/java/org/example/CustomerPages/PageRemoveFromCart.java
@@ -15,18 +15,14 @@ public class PageRemoveFromCart extends Page {
     @Override
     protected void display() throws SQLException {
         Statement isCartEmpty = connection.createStatement();
-        ResultSet customerCartContent =  isCartEmpty.executeQuery("SELECT CustomerID FROM Carts" +
-                " WHERE CustomerID = " + Main.currentCustomerID);
+        ResultSet customerCartContent = isCartEmpty.executeQuery("SELECT CustomerID FROM Carts WHERE CustomerID = " + Main.currentCustomer.ID());
         if (!customerCartContent.next()) {
             System.out.println("Your cart is empty");
             return;
         }
 
-
         Statement getCart = connection.createStatement();
-        ResultSet cart = getCart.executeQuery("SELECT Carts.ID, Carts.Amount, Products.ItemName, Products.Price FROM Carts" +
-                " INNER JOIN Products ON Carts.ProductID = Products.ID" +
-                " WHERE Carts.CustomerID =" + Main.currentCustomerID
+        ResultSet cart = getCart.executeQuery("SELECT Carts.ID, Carts.Amount, Products.ItemName, Products.Price FROM Carts LEFT JOIN Products ON Carts.ProductID = Products.ID WHERE Carts.CustomerID =" +Main.currentCustomer.ID()
         );
 
         System.out.println("0. Back");
@@ -37,7 +33,7 @@ public class PageRemoveFromCart extends Page {
             int price = cart.getInt("Price");
             int amount = cart.getInt("Amount");
 
-            System.out.println(cartIDs.size() - 1 + ": " + productName + ": " + amount + " Price For Each: " + price + ",-");
+            System.out.println(cartIDs.size() + ": " + productName + ": " + amount + " Price For Each: " + price + ",-");
         }
         System.out.println("What item would you like to remove?");
     }
@@ -50,25 +46,27 @@ public class PageRemoveFromCart extends Page {
     @Override
     protected void act() throws SQLException {
         int itemToCancelID = decision();
-        Statement GetInfoOfCancel = connection.createStatement();
-        ResultSet itemToCancel = GetInfoOfCancel.executeQuery("SELECT * FROM Carts WHERE ID =" + cartIDs.get(itemToCancelID));
+        if  (itemToCancelID != 0) {
+            Statement GetInfoOfCancel = connection.createStatement();
+            ResultSet itemToCancel = GetInfoOfCancel.executeQuery("SELECT * FROM Carts WHERE ID =" + cartIDs.get(itemToCancelID - 1));
 
-        int ID = itemToCancel.getInt("ID");
-        int productID = itemToCancel.getInt("ProductID");
-        int amount = itemToCancel.getInt("Amount");
+            int ID = itemToCancel.getInt("ID");
+            int productID = itemToCancel.getInt("ProductID");
+            int amount = itemToCancel.getInt("Amount");
 
 
-        //Add product to stock
-        Statement AddToStock  = connection.createStatement();
-        AddToStock.executeUpdate("UPDATE Products SET Amount = Amount + " + amount + " WHERE ID =" + productID);
+            //Add product to stock
+            Statement AddToStock  = connection.createStatement();
+            AddToStock.executeUpdate("UPDATE Products SET Amount = Amount + " + amount + " WHERE ID =" + productID);
 
-        //Remove from cart
-        Statement RemoveFromCart  = connection.createStatement();
-        RemoveFromCart.executeUpdate("DELETE FROM Carts WHERE ID =" + ID);
-        System.out.println("Product has been removed from your cart");
+            //Remove from cart
+            Statement RemoveFromCart  = connection.createStatement();
+            RemoveFromCart.executeUpdate("DELETE FROM Carts WHERE ID =" + ID);
+            System.out.println("Product has been removed from your cart");
+        }
     }
     private int decision(){
-        return Utils.reader(0,amountOfItemsInCart - 1);
+        return Utils.reader(0,amountOfItemsInCart);
     }
 
 }
